#OSIS-1600 Ajout de la comparaison pour les champs learning_container_year
