+++ base/locale/fr_BE/LC_MESSAGES/django.po
@@ -3396,3 +3396,9 @@ msgstr "Procédure"
 
 msgid "No difference"
 msgstr "Aucune différence"
+
+msgid "is_vacant"
+msgstr "Vacant"
+
+msgid "team"
+msgstr "Gestion par équipe"
