+++ base/templates/learning_unit/comparison.html
@@ -56,7 +56,7 @@
                         {% endfor %}
                         {% if entity_changes %}
                             {% for k in entity_changes %}
-                                <tr {% if forloop.first%}style="border-top: dashed"{% endif %}>
+                                <tr {% if forloop.first%}style="border-top: groove"{% endif %}>
                                     <td style="background-color: lightgray">{% trans k %}</td>
                                     <td>{% value_label entity_changes key=k sub_key='prev' key_comp='current' %}</td>
                                     <td>{% value_label entity_changes key=k sub_key='current' key_comp='current' %}</td>
@@ -64,6 +64,14 @@
                                 </tr>
                             {% endfor %}
                         {% endif %}
+                        {% for key in  fields_lcy %}
+                            <tr {% if forloop.first%}style="border-top: groove"{% endif %}>
+                                <td style="background-color: lightgray">{%  trans key %}</td>
+                                <td>{% if previous_lcy_values|get_item:key %}<label>{{ previous_lcy_values|get_item:key }}</label>{% else %}{{ learning_unit_year.learning_container_year|get_learning_unit_yr_attribute:key | default_if_none:''}}{% endif %}</td>
+                                <td>{{ learning_unit_year.learning_container_year|get_learning_unit_yr_attribute:key | default_if_none:'' }}</td>
+                                <td>{% if next_lcy_values|get_item:key %}<label>{{ next_lcy_values|get_item:key}}</label>{% else %}{{ learning_unit_year.learning_container_year|get_learning_unit_yr_attribute:key | default_if_none:''}}{% endif %}</td>
+                            </tr>
+                        {% endfor%}
                     </tbody>
                 </table>
             </div>
