+++ base/templatetags/learning_unit.py
@@ -29,6 +29,7 @@ from django.utils.translation import ugettext_lazy as _
 
 from base.models.learning_unit_year import find_lt_learning_unit_year_with_different_acronym
 from base.models.proposal_learning_unit import ProposalLearningUnit
+from base.business.learning_units.comparison import DEFAULT_VALUE_FOR_NONE
 
 register = template.Library()
 
@@ -144,11 +145,11 @@ def value_label(values_dict, key, sub_key, key_comp):
     if data:
         val = data.get(sub_key)
         return _get_label(data, key_comp, val)
-    return ''
+    return DEFAULT_VALUE_FOR_NONE
 
 
 def _get_label(data, key_comp, val):
     if val != data.get(key_comp):
-        return mark_safe("<label>{}</label>".format('-' if val is None else val))
+        return mark_safe("<label>{}</label>".format(DEFAULT_VALUE_FOR_NONE if val is None else val))
     else:
-        return mark_safe("{}".format('-' if val is None else val))
+        return mark_safe("{}".format(DEFAULT_VALUE_FOR_NONE if val is None else val))
