+++ base/templatetags/model_field.py
@@ -25,9 +25,9 @@
 ##############################################################################
 from django import template
 from django.utils.translation import ugettext_lazy as _
-from base.models.learning_unit_year import LearningUnitYear, get_value
+from base.models.learning_unit_year import LearningUnitYear
 from django.forms import model_to_dict
-from base.models.learning_unit_year import get_value
+from base.business.learning_units.comparison import get_value
 
 register = template.Library()
 
@@ -35,4 +35,4 @@ register = template.Library()
 @register.filter
 def get_learning_unit_yr_attribute(obj, field_name):
     data = model_to_dict(obj, fields=[field_name])
-    return get_value(LearningUnitYear, data, field_name)
+    return get_value(type(obj), data, field_name)
