+++ base/tests/business/learning_units/test_comparison.py
@@ -27,14 +27,20 @@
 from django.test import TestCase
 from django.test.utils import override_settings
 
-from base.models.learning_unit_year import LearningUnitYear, get_value
-from base.business.learning_units.comparison import get_keys
+from base.models.learning_unit_year import LearningUnitYear
+from base.business.learning_units.comparison import get_keys, _get_changed_values, get_value, \
+    compare_learning_unit_years
+from base.tests.factories.learning_unit_year import create_learning_unit_year
+from base.models.enums import quadrimesters, learning_unit_year_session, attribution_procedure
+from base.models.enums import learning_unit_year_periodicity
+from base.models.enums import learning_unit_year_subtypes
 from base.tests.factories.academic_year import create_current_academic_year, AcademicYearFactory
 from base.tests.factories.learning_unit import LearningUnitFactory
-from base.tests.factories.learning_unit_year import create_learning_unit_year
+from base.tests.factories.learning_unit_year import LearningUnitYearFactory
 
 TITLE = 'Intitulé'
 OTHER_TITLE = 'title 1'
+NEW_ACRONYM = "LDROI1005"
 
 
 class TestComparison(TestCase):
@@ -49,7 +55,7 @@ class TestComparison(TestCase):
     @override_settings(LANGUAGES=[('fr-be', 'French'), ('en', 'English'), ], LANGUAGE_CODE='fr-be')
     def test_get_value_for_boolean(self):
         data = self.learning_unit_year.__dict__
-        self.assertEqual(get_value(LearningUnitYear, data, 'status'), 'Oui')
+        self.assertEqual(get_value(LearningUnitYear, data, 'status'), 'Actif')
 
     @override_settings(LANGUAGES=[('fr-be', 'French'), ('en', 'English'), ], LANGUAGE_CODE='fr-be')
     def test_get_value_for_enum(self):
@@ -65,3 +71,44 @@ class TestComparison(TestCase):
         self.assertCountEqual(get_keys(['a1', 'c3'], ['a1', 'b2', 'c1']), ['a1', 'b2', 'c1', 'c3'])
 
 
+class LearningUnitYearComparaisonTest(TestCase):
+
+    def setUp(self):
+        self.academic_year = create_current_academic_year()
+        self.previous_academic_year = AcademicYearFactory(year=self.academic_year.year + 1)
+
+        self.learning_unit_year = self.create_learning_unit_yr(self.academic_year, "LDR", True,
+                                                               learning_unit_year_subtypes.FULL)
+        self.previous_learning_unit_year = self.create_learning_unit_yr(self.previous_academic_year, NEW_ACRONYM, False,
+                                                                        learning_unit_year_subtypes.PARTIM)
+
+    def create_learning_unit_yr(self, academic_year, acronym, status, subtype):
+        return LearningUnitYearFactory(acronym=acronym,
+                                       academic_year=academic_year,
+                                       subtype=subtype,
+                                       status=status,
+                                       internship_subtype=None,
+                                       credits=5,
+                                       periodicity=learning_unit_year_periodicity.ANNUAL,
+                                       language=None,
+                                       professional_integration=True,
+                                       specific_title="Juridic law courses",
+                                       specific_title_english=None,
+                                       quadrimester=quadrimesters.Q1,
+                                       session=learning_unit_year_session.SESSION_123,
+                                       attribution_procedure=attribution_procedure.EXTERNAL
+                                       )
+
+    @override_settings(LANGUAGES=[('fr-be', 'French'), ('en', 'English'), ], LANGUAGE_CODE='fr-be')
+    def test_compare(self):
+        self.assertCountEqual(compare_learning_unit_years(self.learning_unit_year, self.previous_learning_unit_year),
+                              {'acronym': NEW_ACRONYM,
+                               'status': 'Non',
+                               'subtype': 'Partim'})
+
+    @override_settings(LANGUAGES=[('fr-be', 'French'), ('en', 'English'), ], LANGUAGE_CODE='fr-be')
+    def test_get_changed_value_wrong_fieldname(self):
+        data_obj1, data_obj2 = self.learning_unit_year.__dict__, self.previous_learning_unit_year.__dict__
+        del data_obj2['acronym']
+        with self.assertRaises(KeyError):
+            _get_changed_values(data_obj1, data_obj2, ['acronym'], LearningUnitYear)
