+++ base/views/learning_unit.py
@@ -39,7 +39,8 @@ from base.business.learning_unit import get_cms_label_data, \
     get_same_container_year_components, find_language_in_settings, \
     CMS_LABEL_SPECIFICATIONS, get_achievements_group_by_language
 from base.business.learning_units import perms as business_perms
-from base.business.learning_units.comparison import get_keys
+from base.business.learning_units.comparison import get_keys, compare_learning_unit_years, \
+    compare_learning_container_years
 from base.business.learning_units.perms import can_update_learning_achievement
 from base.forms.learning_class import LearningClassEditForm
 from base.forms.learning_unit_component import LearningUnitComponentEditForm
@@ -247,14 +248,21 @@ def learning_unit_comparison(request, learning_unit_year_id):
     previous_academic_yr = mdl.academic_year.find_academic_year_by_year(learning_unit_yr.academic_year.year - 1)
     previous_lu = get_learning_unit(previous_academic_yr, learning_unit_yr)
     previous_values = compare_learning_unit(learning_unit_yr, previous_lu)
+    previous_lcy_values = compare_learning_container_years(learning_unit_yr.learning_container_year,
+                                                           previous_lu.learning_container_year)
 
     next_academic_yr = mdl.academic_year.find_academic_year_by_year(learning_unit_yr.academic_year.year + 1)
     next_lu = get_learning_unit(next_academic_yr, learning_unit_yr)
     next_values = compare_learning_unit(learning_unit_yr, next_lu)
+    next_lcy_values = compare_learning_container_years(learning_unit_yr.learning_container_year,
+                                                       next_lu.learning_container_year)
 
     previous_context = get_learning_unit_comparison_context(previous_lu)
     next_context = get_learning_unit_comparison_context(next_lu)
     entity_changes = _get_changed_organization(context, previous_context, next_context)
+    # learning_container_yr_changes = _get_learning_container_yr_changes(context['learning_unit_year'],
+    #                                                                    previous_context['learning_unit_year'],
+    #                                                                    next_context['learning_unit_year'])
 
     context.update(
         {'previous_values': previous_values,
@@ -263,12 +271,15 @@ def learning_unit_comparison(request, learning_unit_year_id):
          'next_values': next_values,
          'fields': get_keys(list(previous_values.keys()), list(next_values.keys())),
          'entity_changes': entity_changes,
+         'fields_lcy': get_keys(list(previous_lcy_values.keys()), list(next_lcy_values.keys())),
+         'previous_lcy_values': previous_lcy_values,
+         'next_lcy_values': next_lcy_values
          })
     return layout.render(request, "learning_unit/comparison.html", context)
 
 
 def compare_learning_unit(learning_unit_yr, learning_unit_yr_other):
-    return learning_unit_yr.compare(learning_unit_yr_other)
+    return compare_learning_unit_years(learning_unit_yr, learning_unit_yr_other)
 
 
 def get_learning_unit(academic_yr, learning_unit_yr):
